move test positions into more logical packages
