+++ src/front_end/ui/DisplayTaskListUI.java
@@ -1,8 +1,6 @@
 package front_end.ui;
 
 import back_end.storage.relations.Task;
-import front_end.VisualTuple;
-import front_end.ui.base.UserInterface;
 import front_end.ui.base.VisualIndexUI;
 
 import java.util.List;
